course controllers updated

+++ src/controllers/course.controller.js
@@ -6,17 +6,19 @@ import { ApiResponse } from "../utils/ApiResponse.js";
 import { asyncHandler } from "../utils/asyncHandler.js";
 import { deleteMediaFromCloudinary, uploadOnCloudinary, deleteVideoFromCloudinary } from "../utils/cloudinary.js";
 
+
 const createCourse = asyncHandler(async (req, res) => {
-  const { courseTitle, category } = req.body;
-  if (!courseTitle || !category) {
+  const { courseTitle, courseDescription, coursePrice } = req.body;
+  if (!courseTitle || !courseDescription || !coursePrice) {
     throw new ApiError(400, "all fields are required")
   }
 
   try {
     const course = await Course.create({
       courseTitle,
-      category,
-      creator: req.user._id,
+      courseDescription,
+      coursePrice,
+      instructorId: req.user?._id,
     })
 
     return res
@@ -38,7 +40,7 @@ const searchCourse = asyncHandler(async (req, res) => {
     const { query = "", categories = [], sortByPrice } = req.query;
 
     const searchCriteria = {
-      ispublished: true,
+      isPublished: true,
       $or: [
         { courseTitle: { $regex: query, $options: "i" } },
         { subTitle: { $regex: query, $options: "i" } },
@@ -57,10 +59,8 @@ const searchCourse = asyncHandler(async (req, res) => {
       sortOptions.coursePrice = -1;//sort by price in ascending
     }
 
-    const courses = await Course.find(searchCriteria).populate({
-      path: "creator",
-      select: "name photoUrl"
-    }).sort(sortOptions)
+    const courses = await Course.find(searchCriteria).populate({path: "instructorId", select: "name photoUrl"}).sort(sortOptions)
+
 
     return res
       .status(200)
@@ -77,10 +77,9 @@ const searchCourse = asyncHandler(async (req, res) => {
   }
 })
 
-// by course creator
 const getpublishedCourse = asyncHandler(async (req, res) => {
   try {
-    const courses = await Course.find({ ispublished: true }).populate({ path: "creator", select: "name photoUrl" })
+    const courses = await Course.find({ isPublished: true }).populate({ path: "instructorId", select: "name photoUrl" })
 
     if (!courses) {
       return res.status(404).json({
@@ -98,11 +97,10 @@ const getpublishedCourse = asyncHandler(async (req, res) => {
   }
 })
 
-// get creator course by creator
-const getCreatorCourses = asyncHandler(async (req, res) => {
-  const userId = req.user._id;
+const getInstructorCourses = asyncHandler(async (req, res) => { 
+  const instructorId = req.user._id;
   try {
-    const courses = await Course.find({ creator: userId })
+    const courses = await Course.find({ instructorId })
     if (!courses) {
       throw new ApiError(400, "courses not found")
     }
@@ -112,18 +110,18 @@ const getCreatorCourses = asyncHandler(async (req, res) => {
       .json(
         new ApiResponse(
           200,
-          { courses: [] },
-          "get creator course found successfully"
+          { courses },
+          "get instructor course found successfully"
         )
       )
   } catch (error) {
-    throw new ApiError(500, error?.message, "error getting creator courses")
+    throw new ApiError(500, error?.message, "error getting instructor courses")
   }
 })
 
 const editCourse = asyncHandler(async (req, res) => {
-  const courseId = req.params.courseId;
-  const { courseTitle, subTitle, description, category, courseLevel, coursePrice } = req.body;
+  const {courseId} = req.params;
+  const { courseTitle, courseDescription, coursePrice } = req.body;
   const thumbnailLocalPath = req.files?.thumbnail[0].path;
 
   try {
@@ -142,7 +140,7 @@ const editCourse = asyncHandler(async (req, res) => {
       courseThumbnail = await uploadOnCloudinary(thumbnailLocalPath)
     }
 
-    const updateData = { courseTitle, subTitle, description, category, courseLevel, coursePrice, courseThumbnail: courseThumbnail?.path }
+    const updateData = { courseTitle, courseDescription, coursePrice, courseThumbnail: courseThumbnail?.url }
 
     course = await Course.findByIdAndUpdate(
       courseId,
@@ -201,7 +199,7 @@ const createLecture = asyncHandler(async (req, res) => {
     throw new ApiError(400, "invalid course id");
 
   if (!lectureTitle || !courseId) {
-    throw new ApiError(400, "invalid course id");
+    throw new ApiError(400, "lecture title and course id is required");
   };
 
   try {
@@ -252,7 +250,7 @@ const getCourseLecture = asyncHandler(async (req, res) => {
       )
 
   } catch (error) {
-
+    throw new ApiError(500, error?.message, "error getting course lecture")
   }
 })
 
@@ -265,9 +263,9 @@ const editLecture = asyncHandler(async (req, res) => {
     throw new ApiError(400, "invalid course id or lecture id");
 
   try {
-    const lecture = await Lecture.findById(lectureId);
+    let lecture = await Lecture.findById(lectureId);
     if (!lecture) {
-      throw new ApiError(400, "invalid course id");
+      throw new ApiError(400, "invalid lecture id");
     }
 
     if (lectureTitle) lecture.lectureTitle = lectureTitle;
@@ -299,50 +297,50 @@ const editLecture = asyncHandler(async (req, res) => {
 })
 
 const removeLecture = asyncHandler(async (req, res) => {
-  try {
-    const { lectureId } = req.params;
-    if (!isValidObjectId(lectureId))
-      throw new ApiError(400, "invalid lecture id");
+  const { courseId, lectureId } = req.params;
 
-    const lecture = await Lecture.findByIdAndDelete(lectureId);
-    if (!lecture) {
-      return res.status(404).json({
-        message: "Lecture not found!"
-      });
+  if (!isValidObjectId(courseId) || !isValidObjectId(lectureId))
+    throw new ApiError(400, "invalid course id or lecture id");
+
+  try {
+    let course = await Course.findById(courseId);
+    if (!course) {
+      throw new ApiError(400, "invalid course id");
     }
+
+    const lecture = await Lecture.findById(lectureId);
+    if (!lecture) {
+      throw new ApiError(400, "invalid lecture id");
+    }
+
+    await Lecture.findByIdAndDelete(lectureId);
+
+    course.lectures = course.lectures.filter((lec) => lec.toString() !== lectureId.toString());
+    await course.save({ validateBeforeSave: false });
 
     if (lecture.publicId) {
       await deleteVideoFromCloudinary(lecture.publicId);
     }
-
-    await Course.updateOne(
-      { lectures: lectureId },
-      { $pull: { lectures: lectureId } }
-    )
-
 
     return res
       .status(200)
       .json(
         new ApiResponse(
           200,
-          {},
-          "lecture removed successfully"
+          { message: "lecture deleted successfully" },
+          "lecture deleted successfully"
         )
       )
 
   } catch (error) {
-    console.log(error);
-    return res.status(500).json({
-      message: "Failed to remove lecture"
-    })
+    throw new ApiError(400, error?.message, "invalid course id");
   }
 })
 
 export {
   createCourse,
   searchCourse,
-  getCreatorCourses,
+  getInstructorCourses,
   getpublishedCourse,
   editCourse,
   getCourseById,
